RISC-V: Remove the current perf implementation
The current perf implementation in RISC-V is not very useful as it can not
count any events other than cycle/instructions. Moreover, perf record
can not be used or the events can not be started or stopped.

Remove the implementation now for a better platform driver in future
that will implement most of the missing functionality.

Reviewed-by: Anup Patel <anup@brainfault.org>
Signed-off-by: Atish Patra <atish.patra@wdc.com>
Signed-off-by: Atish Patra <atishp@rivosinc.com>

Rebase note: Resolved conflict with
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/arch/riscv/kernel/Makefile?h=linux-5.15.y&id=1851b9a467065b18ec2cba156eea345206df1c8f
leading trace_irq, removing perf_event

Rebase note: Resolved conflict with
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/arch/riscv/Kconfig?h=linux-5.15.y&id=9c7ee9471550a7fd6f8665c6ec622b0284244786
as of adjacent line change

Signed-off-by: MichaIng <micha@dietpi.com>